Look at the description
NOTE to whoever is looking at this: the functions "by-x" are all basically the 
same code, this was initially to make it so that it's easier to edit each individual functions.
Still have to do number of cancellations --> which will likely be done by a different function compared to the by-x functions. 
And I also have to work on the ATA number sorting --> again, another function.