Read description
So, I have tidied up a bit. ATA repairing is now included in the genPickle which is the renamed file that generates the pickle from the csv. All pure math functions so far (least squ, rank) are now in numericals.py. Look at the TrivialStats.py for examples how to use the functions. Also, trivialstats makes a start on some very simple data analysis